LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/multicol-gap-animation-003-expected.txt


FAIL Default column-gap is not interpolable assert_equals: expected "100px" but got "80.24139404296875px"
